mobile: minor refactor, and ui improvements
abstract layout carousel component
