+++ mobile-app/src/components/app-maker/index.ts
@@ -1,5 +1,6 @@
 export { AnimatedFlingContainer } from './AnimatedFlingContainer';
 export { LayoutDivider } from './LayoutDivider';
+export { LayoutCarousel } from './LayoutCarousel';
 export { MakerConfiguration } from './maker-configuration';
 export { AppMakerScreenAction } from './AppMakerScreenAction';
 export { MakerBox } from './MakerBox';
